requirements.txt
wpsycopg2~=2.9.1